requirements.txt
billiard==4.1.0
black==23.7.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
amqp==5.1.1
asgiref==3.7.2
Django==4.1.10
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
kombu==5.3.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
